Update location of initial states and store
