+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './components/App';
-import store from './store';
+import store from './reducers/index';
 
 ReactDOM.render(
   <React.StrictMode>
