+++ src/reducers/index.js
@@ -1,8 +1,21 @@
-import { combineReducers } from 'redux';
-import bookReducer from './books';
+import { createStore } from 'redux';
+import rootReducer from './rootReducer';
 
-const rootReducer = combineReducers({
-  book: bookReducer,
-});
+const initialState = {
+  book: [
+    {
+      id: 1,
+      title: 'MacBeth',
+      category: 'History',
+    },
+    {
+      id: 2,
+      title: 'Rome and Juliet',
+      category: 'Romantic',
+    },
+  ],
+};
 
-export default rootReducer;
+const store = createStore(rootReducer, initialState);
+
+export default store;
