+++ src/reducers/rootReducer.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import bookReducer from './books';
+
+const rootReducer = combineReducers({
+  book: bookReducer,
+});
+
+export default rootReducer;
